Aitor->Master

+++ web/controller/PedidoController.php
@@ -160,24 +160,31 @@ class PedidoController extends Controller
 
     public function confirmarPedido($id)
     {
-        if(isset($_POST['id'])){
-            $id = $_POST['id'];
-        }
-        $pedido = new Pedido($this->conexion);
-        $pedido->updateEstado($id, 1);
+        if (parent::verifyAdmin()){
+            if(isset($_POST['id'])){
+                $id = $_POST['id'];
+            }
+            $pedido = new Pedido($this->conexion);
+            $pedido->updateEstado($id, 1);
 //        header('Location: /index.php?controller=pedido');
-        die();
+            die();
+        }else{
+            header("Location: index.php?controller=producto");
+        }
     }
     public function rechazarPedido($id)
     {
-        if(isset($_POST['id'])){
-            $id = $_POST['id'];
-        }
-        $p = new Pedido($this->conexion);
-        $pedido = $p->deleteByID($id);
+        if (parent::verifyAdmin()){
+            if(isset($_POST['id'])){
+                $id = $_POST['id'];
+            }
+            $p = new Pedido($this->conexion);
+            $pedido = $p->deleteByID($id);
 //        header('Location: /index.php?controller=pedido');
-        die();
-
+            die();
+        }else{
+            header("Location: index.php?controller=producto");
+        }
     }
 
     public function mostrarCarrito()
